requirements.txt
# etl
wget

# spark
pyspark
graphframes-py
typing-extensions
networkx
matplotlib
